MovingEddies: Make grid generation a function and move to test
